Upgrade register/unregister, fix bugs

+++ api/dbmodels/client.py
@@ -1,5 +1,6 @@
 from typing import List
 
+import discord
 from sqlalchemy.ext.hybrid import hybrid_property
 
 from api.database import db
@@ -11,7 +12,7 @@ class Client(db.Model, Serializer):
 
     # Identification
     id = db.Column(db.Integer, primary_key=True)
-    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=True)
+    user_id = db.Column(db.BigInteger, db.ForeignKey('user.id'), nullable=True)
     discord_user_id = db.Column(db.Integer, db.ForeignKey('discorduser.id'), nullable=True)
 
     # User Information
